thêm count follow, thumbnail cho vol

+++ src/main/java/com/example/readnovel/entity/Chapter.java
@@ -22,7 +22,7 @@ public class Chapter extends BaseEntity {
     private String id;
     private String title;
     private String numberTitle;
-    private float number;
+    private double number;
     @Column(columnDefinition = "TEXT")
     private String content;
     @ManyToOne()
